fix: force fp32 for A in _build_real_A (selective_scan_fn dtype requirement)
selective_scan_fn requires A in fp32. When model is cast to bf16 via
model.to(torch.bfloat16), self.nu becomes bf16, making decay/A_diag/A
all bf16 — crashing the kernel. Fix: .float() on nu before softplus and
on the final A tensor. D and delta_bias already had .float() in the call.

+++ model/arg_layer.py
@@ -176,11 +176,9 @@ class ARGLayer(nn.Module):
         internally via exp(A * Δ), so we pass the continuous-time log-decay
         (negative real, per state channel).
         """
-        decay = -F.softplus(self.nu)          # (d_state_complex,) — negative continuous decay
-        # Each complex eigenvalue → 2 real state channels with the same decay
+        decay = -F.softplus(self.nu.float())   # fp32 — mamba_ssm requires A in fp32
         A_diag = decay.repeat_interleave(2)   # (d_state,)
-        # Broadcast across d_inner channels: shape (d_inner, d_state)
-        A = A_diag.unsqueeze(0).expand(self.d_inner, -1).contiguous()
+        A = A_diag.unsqueeze(0).expand(self.d_inner, -1).contiguous().float()  # (d_inner, d_state), fp32
         return A
     
     # ─────────────────────────────────────────────────────────────────────────
